fix wantParamsWrapper

AI[100%] Human Fixed[0%] Human[0%] AI Adopted[100%]

+++ test/unittest/want/want_params_wrapper_test.cpp
@@ -905,7 +905,7 @@ HWTEST_F(WantParamWrapperBaseTest, Want_Param_Wrapper_4200, Function | MediumTes
 HWTEST_F(WantParamWrapperBaseTest, Want_Param_Wrapper_4300, Function | MediumTest | Level1)
 {
     WantParams params;
-    params.SetParam("array", sptr<Array>::MakeSptr(0, g_IID_IInteger));
+    params.SetParam("array", sptr<Array>::MakeSptr(static_cast<long>(0), g_IID_IInteger));
     WantParamWrapper wrapper(params);
 
     std::string serialized = wrapper.ToString();
